refactor(chat-hooks): extract resume-attach flow and buffered-generation helpers
useConversationStream.ts had grown to ~800 lines, mostly from
resumeIfAwaitingGeneration's runWatch/runAttach — two near-identical SSE
line-parsing loops. Extracts:

- buffered-generation.ts: BufferedGeneration + restoreBufferedMessage,
  shared by startStream's live-chunk path and the resume path.
- generation-resume.ts: the whole resume mechanism as
  createResumeIfAwaitingGeneration(deps), plus a shared readSseEvents helper
  so runWatch/runAttach only supply their own event handler instead of each
  reimplementing the read loop.

useConversationStream.ts now composes the extracted resumer via useMemo
(useCallback requires an inline function literal for its dependency
analysis, which a factory call doesn't satisfy). No behavior change — same
tests, unchanged.
